+++ Model/ResourceModel/Config/Collection.php
@@ -26,7 +26,7 @@ class Collection extends AbstractCollection
     /**
      * @inheritDoc
      */
-    protected function _construct()
+    protected function _construct(): void
     {
         $this->_init(Config::class, ResourceModel\Config::class);
     }